refactor : console 애플리케이션 페키지 분리
