+++ src/main/java/team/marco/voucher_management_system/console_app/application/CommandCustomerApplication.java
@@ -1,4 +1,4 @@
-package team.marco.voucher_management_system.application;
+package team.marco.voucher_management_system.console_app.application;
 
 import java.util.NoSuchElementException;
 import org.slf4j.Logger;
@@ -6,7 +6,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.jdbc.datasource.lookup.DataSourceLookupFailureException;
 import org.springframework.stereotype.Component;
-import team.marco.voucher_management_system.controller.ConsoleCustomerController;
+import team.marco.voucher_management_system.console_app.controller.ConsoleCustomerController;
 import team.marco.voucher_management_system.type_enum.CustomerCommandType;
 import team.marco.voucher_management_system.util.Console;
