+++ src/main/java/team/marco/voucher_management_system/console_app/controller/ConsoleVoucherController.java
@@ -1,13 +1,13 @@
-package team.marco.voucher_management_system.controller;
+package team.marco.voucher_management_system.console_app.controller;
 
 import java.text.MessageFormat;
 import java.util.List;
-import org.springframework.stereotype.Controller;
+import org.springframework.stereotype.Component;
 import team.marco.voucher_management_system.model.Voucher;
 import team.marco.voucher_management_system.service.VoucherService;
 import team.marco.voucher_management_system.util.Console;
 
-@Controller
+@Component
 public class ConsoleVoucherController {
     private static final String INFO_DELIMINATOR = MessageFormat.format("\n{0}\n", "-".repeat(42));
